Reflect different classification details, split file
