+++ js_src/coaching/PlanView.js
@@ -1,155 +1,9 @@
 import React, { useState } from 'react';
 
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { regular } from '@fortawesome/fontawesome-svg-core/import.macro';
-
-import store from 'utils/store';
+import Explainer from 'utils/Explainer';
 import Loader from 'utils/Loader';
-
-
-const TierEvent = ({ ev, planId }) => {
-    let example = {"name":"Good Friday 720","eventFormat":"Single 720","ageGroups":"U21 & U18 only","date":{"api":"2023-04-07","pretty":"Friday 7 April"},"venue":{"name":"Wallingford Sports Park","postCode":"OX10 9RB"},"tournament":null,"clubEvent":null,"clubTrip":false,"entryLink":null,"registration":null};
-
-    let eventType = null;
-    if (ev.tournament) {
-        eventType = 'WCA hosted competition';
-    } else if (ev.clubEvent) {
-        eventType = 'WCA hosted informal';
-    } else if (ev.clubTrip) {
-        eventType = 'WCA club trip';
-    } else {
-        eventType = 'DIY event';
-    }
-
-    let options = (
-        <>
-            <option value="">{ ev.registration && 'Delete response' || 'Respond...' }</option>
-            <option value="definite">Yes, I am definitely attending</option>
-            <option value="booked">Yes, I am booked in</option>
-            <option value="maybe">I might be attending</option>
-            <option value="no">I will not be attending</option>
-        </>
-    );
-    let defaultValue = ev.registration && ev.registration.status.id;
-    if (ev.clubTrip) {
-        options = (
-            <>
-                <option value="">{ ev.registration && 'Delete response' || 'Respond...' }</option>
-                <option value="definite">Yes, I am definitely attending</option>
-                <option value="booked|required">Booked, I need transport</option>
-                <option value="booked|interested">Booked, I would like transport for me</option>
-                <option value="booked|plus-parent">Booked, I would like transport for me and a parent</option>
-                <option value="booked|own-way">Booked, I do not need transport</option>
-                <option value="maybe">I might be attending</option>
-                <option value="no">I will not be attending</option>
-            </>
-        );
-        if (ev.registration && ev.registration.status.id === 'booked' && ev.registration.wantsTransport) {
-            defaultValue = `booked|${ev.registration.wantsTransport.id}`;
-        }
-    }
-
-    let link = null;
-    let linkText = null;
-    if (ev.tournament) {
-        link = ev.tournament.link;
-        linkText = 'Book via club website';
-    } else if (ev.clubEvent) {
-        link = ev.clubEvent.link;
-        linkText = 'Book via club website';
-    } else if (ev.entryLink) {
-        link = ev.entryLink;
-        linkText = 'External booking';
-    }
-
-    const register = (e) => {
-        const data = {
-            status: e.target.value,
-        };
-        if (data.status.includes('|')) {
-            const [stat, transport] = data.status.split('|');
-            data.status = stat;
-            data.wantsTransport = transport;
-        }
-        store.send(ev.registrationLink, data).then((response) => {
-            if (response.ok) {
-                store.invalidate(`/api/coaching/plan/${planId}/`);
-            }
-        }).catch((error) => {
-            console.error('error', error);
-        });
-    };
-
-    return (
-        <div className="event-plan__event">
-            <h4 className="event-plan__event__title">
-                { ev.name }
-                { ev.registration &&
-                    <span>
-                        <FontAwesomeIcon icon={ regular('circle-check') } className="event-plan__icon" />
-                        { ev.registration.status.display } 
-                    </span>
-                }
-            </h4>
-            <p className="event-plan__event__date">
-                <FontAwesomeIcon icon={ regular('calendar-days') } className="event-plan__icon" fixedWidth />
-                { ev.date.pretty }
-                { ev.endDate && <> - { ev.endDate.pretty }</>}
-            </p>
-            <p className="event-plan__event__location">
-                <FontAwesomeIcon icon={ regular('map') } className="event-plan__icon" fixedWidth />
-                { ev.venue.name }, { ev.venue.postCode }
-            </p>
-            <p className="event-plan__event__format">
-                <FontAwesomeIcon icon={ regular('rectangle-list') } className="event-plan__icon" fixedWidth />
-                { eventType } - { ev.eventFormat }
-            </p>
-            <p className="event-plan__event__format">
-                <FontAwesomeIcon icon={ regular('user') } className="event-plan__icon" fixedWidth />
-                { ev.ageGroups }
-            </p>
-            <select className="event-plan__event__registration" defaultValue={ defaultValue } onChange={ register }>
-                { options }
-            </select>
-            { link && <a className="event-plan__event__booking" href={ link }>{ linkText }</a> }
-        </div>
-    );
-};
-
-
-const Tier = ({ tier, planId }) => {
-    const events = tier.events.map(ev => <TierEvent ev={ ev } key={ `${ev.date.api} / ${ev.name}` } planId={ planId } />);
-    return (
-        <div className="event-plan__tier">
-            <h3 className="event-plan__heading">{ tier.tierName } - { tier.name }</h3>
-            <p className="event-plan__content">{ tier.comments }</p>
-            <div className="event-plan__events">
-                { events }
-            </div>
-        </div>
-    );
-};
-
-
-const Explainer = ({ title, children }) => {
-    const [open, setOpen] = useState(false);
-    let cls = 'explainer';
-    if (open) {
-        cls += ' explainer--open';
-    }
-    return (
-        <div className={ cls }>
-            <h3 className="explainer__question" onClick={ () => setOpen(!open) }>{ title }</h3>
-            { open && children || null }
-        </div>
-    );
-};
-Explainer.p = ({ children }) => {
-    return <p className="explainer__content">{ children }</p>;
-};
-Explainer.table = ({ children }) => {
-    return <table className="explainer__table">{ children }</table>;
-};
+import ClassificationExplainer from './ClassificationExplainer';
+import Tier from './Tier';
 
 
 class PlanView extends Loader {
@@ -192,20 +46,7 @@ class PlanView extends Loader {
                         </Explainer.p>
                     </Explainer>
 
-                    <Explainer title="How do I achieve a Bowman 1st class classification?">
-                        <Explainer.p>
-            In order to achieve a Bowman 1st class classification, you need to shoot 18
-            dozen arrows (equivalent to 1.5 days of competition) at competitive
-            events, with each event being over the relevant threshold score.
-            Any competitive event is suitable, whether it’s a club target day,
-            club champs, or anything else all the way up to National
-            Championships. The only requirement is that it either includes
-            your longest distance, or is your 720 round.
-                        </Explainer.p>
-                        <Explainer.p>
-            Full classification tables can be found at <a href="https://archerygeekery.co.uk/mobile-friendly-classification-tables">Archery Geekery</a>.
-                        </Explainer.p>
-                    </Explainer>
+                    <ClassificationExplainer classification={ data.targetClassification } />
 
                     <Explainer title="What rounds and distances do I shoot?">
                         <Explainer.p>
